One more place where size was `int`

+++ src/main/java/alien/catalogue/access/XrootDEnvelope.java
@@ -140,7 +140,7 @@ public class XrootDEnvelope implements Serializable {
 			String lfn = "";
 			String guid = "";
 			String se = "";
-			int size = 0;
+			long size = 0;
 			String md5 = "";
 
 			unEncryptedEnvelope = envelope;
@@ -173,7 +173,7 @@ public class XrootDEnvelope implements Serializable {
 					else if ("guid".equals(key))
 						guid = value;
 					else if ("size".equals(key))
-						size = Integer.parseInt(value);
+						size = Long.parseLong(value);
 					else if ("md5".equals(key))
 						md5 = value;
 					else if ("se".equals(key))
